pyproject.toml
[project]
name = "universe-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "grpcio>=1.50.0",
    "grpcio-tools>=1.50.0",
    "numpy>=1.20.0",
    "pip>=25.0.1",
    "pyaudio>=0.2.13",
    "requests>=2.32.3",
    "sounddevice>=0.5.1",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "uvicorn>=0.34.0",
    "websockets>=15.0.1",
]

[build-system]
requires = ["setuptools>=67.0.0"]
build-backend = "setuptools.build_meta"

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
